+++ NEWS.md
@@ -3,6 +3,7 @@
 * Improves and simplifies tests
 * Fixes code to work on cropped stars proxy
 * Adds a type argument to `lsp_add_quality()`, making it possible to calculate either isolation (type = "segmentation") or distinction (type = "cluster")
+* Adds `lsp_add_examples()` to extract and add stars objects to an lsp object
 
 # motif 0.3.12